arc: Introduce base input clk device
This commit intoduces base input clock device for AXS10x
boards. It provides a fixed clock rate of 33,333333 MHz.

It is used as an input clock by uart0 driver, so we add
corresponding reference to uart0 node in device tree.

This commit fixes work of uart driver on Synopsysi AXS10x
boars.